Github issue #27.  Update OznMon to work in GSI-monitor, port to orion.  Completes #27.

+++ src/Ozone_Monitor/image_gen/ush/plot_horiz.sh
@@ -1,16 +1,11 @@
-#! /bin/ksh -l
+#! /bin/bash
 
 #------------------------------------------------------------------
 #  plot_horiz.sh
 #
 
 echo "begin plot_horiz.sh"
-set -ax
 
-
-if [[ ${MY_MACHINE} = "hera" ]]; then
-   module load grads
-fi
 
 export SATYPE=$1
 export PVAR=$2
@@ -19,12 +14,16 @@ export dsrc=$4		# data source -- ges | anl
 
 echo "SATYPE, PVAR, PTYPE dsrc = $SATYPE, $PVAR, $PTYPE, $dsrc"
 echo "RUN = $RUN"
+echo "NDATE = $NDATE"
+echo "PDATE = $PDATE"
 
 #------------------------------------------------------------------
 # Set work space for this SATYPE source.
 #
 tmpdir=${WORKDIR}/${SATYPE}.${dsrc}.$PDATE.${PVAR}
-rm -rf $tmpdir
+if [[ -d $tmpdir ]]; then
+   rm -rf $tmpdir
+fi
 mkdir -p $tmpdir
 cd $tmpdir
 
@@ -41,22 +40,29 @@ while [[ $ctr -le 3 ]]; do
    c_pdy=`echo $cdate|cut -c1-8`
    c_cyc=`echo $cdate|cut -c9-10`
 
-   tankdir_cdate=${TANKDIR}/${RUN}.${c_pdy}/${c_cyc}/atmos/oznmon/horiz
+   tankdir_cdate=${OZN_TANKDIR_STATS}/${RUN}.${c_pdy}/${c_cyc}/atmos/oznmon/horiz
    if [[ ! -d ${tankdir_cdate} ]]; then
-      tankdir_cdate=${TANKDIR}/${RUN}.${c_pdy}/${c_cyc}/oznmon/horiz
+      tankdir_cdate=${OZN_TANKDIR_STATS}/${RUN}.${c_pdy}/${c_cyc}/oznmon/horiz
       if [[ ! -d ${tankdir_cdate} ]]; then
-         tankdir_cdate=${TANKDIR}/${RUN}.${c_pdy}/horiz
+         tankdir_cdate=${OZN_TANKDIR_STATS}/${RUN}.${c_pdy}/horiz
       fi
    fi
 
-   $NCP ${tankdir_cdate}/${SATYPE}.${dsrc}.ctl ./
-   $NCP ${tankdir_cdate}/${SATYPE}.${dsrc}.${c_pdy}* ./
+   if [[ -e ${tankdir_cdate}/${SATYPE}.${dsrc}.ctl ]]; then 
+      $NCP ${tankdir_cdate}/${SATYPE}.${dsrc}.ctl ./
+   fi
+
+   if compgen -G "${tankdir_cdate}/${SATYPE}.${dsrc}.${c_pdy}*ieee_d*" > /dev/null; then
+      $NCP ${tankdir_cdate}/${SATYPE}.${dsrc}.${c_pdy}*ieee_d* ./
+   fi
 
    cdate=`$NDATE -6 $cdate`
    ctr=`expr $ctr + 1`
 done
 
-$UNCOMPRESS *${Z}
+if compgen -G "*.gz" > /dev/null; then
+   $UNCOMPRESS *.gz
+fi
 
 #----------------------------------------------------------------
 #  Modify tdef line in .ctl file to start at bdate.
@@ -64,38 +70,43 @@ $UNCOMPRESS *${Z}
 if [[ -e ${SATYPE}.${dsrc}.ctl ]]; then
    edate=`$NDATE -18 $PDATE`
    ${OZN_IG_SCRIPTS}/update_ctl_tdef.sh ${SATYPE}.${dsrc}.ctl ${edate} 4
-fi
 
-$NCP ${OZN_IG_GSCRPTS}/cbarnew.gs ./
+   $NCP ${OZN_IG_GSCRPTS}/cbarnew.gs ./
 
 
-$STNMAP -i ${SATYPE}.${dsrc}.ctl
+   $STNMAP -i ${SATYPE}.${dsrc}.ctl
 
-for var in ${PTYPE}; do
+   for var in ${PTYPE}; do
 
 cat << EOF > ${SATYPE}_${var}.gs
 'open ${SATYPE}.${dsrc}.ctl'
 'run ${OZN_IG_GSCRPTS}/plot_horiz_${dsrc}.gs ${OZNMON_SUFFIX} ${RUN} ${SATYPE} ${var} x800 y700'
 'quit'
 EOF
-   $GRADS -blc "run ${tmpdir}/${SATYPE}_${var}.gs"   
+      echo "GRADS = $GRADS"
+      $GRADS -blc "run ${tmpdir}/${SATYPE}_${var}.gs"   
 
 
-done 
+   done 
 
 
-#--------------------------------------------------------------------
-#   Move image files to TANK
-#
+   #--------------------------------------------------------------------
+   #   Move image files to TANK
+   #
 
-${NCP} *.png ${OZN_IMGN_TANKDIR}/.
+   ${NCP} *.png ${OZN_IMGS_HORIZ}/.
 
+else
+   echo "Unable to plot $SATYPE, no ctl file was found" 
+fi
 
 #--------------------------------------------------------------------
 # Clean $tmpdir.  Submit done job.
-#cd $tmpdir
-#cd ../
-#rm -rf $tmpdir
+#
+if [[ $KEEPDATA -ne 1 ]]; then
+   cd ../
+   rm -rf $tmpdir
+fi
 
 echo "end plot_horiz.sh"
 exit
